+++ lib/cocoapods_plugins.rb
@@ -2,5 +2,5 @@
 # The namespace of the Cocoapods plugins plugin.
 #
 module CocoapodsPlugins
-  VERSION = '0.3.2'
+  VERSION = '0.4.0'
 end